% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_campaign_vaccination_coverage_file}
\alias{create_campaign_vaccination_coverage_file}
\title{Create a vaccine coverage file for campaign-only scenario}
\usage{
create_campaign_vaccination_coverage_file(
  campaign_only_vaccination_coverage_file,
  routine_campaign_vaccination_coverage_file
)
}
\arguments{
\item{campaign_only_vaccination_coverage_file}{File name given to the
campaign-only scenarios}

\item{routine_campaign_vaccination_coverage_file}{File used to generate
vaccine coverage for campaign-only scenarios}
}
\description{
A csv file for campaign-only vaccination scenario is created based on the
coverage file for (routine + campaign) scenarios, by setting routine coverage
 to zero. This is done because routine coverage values are needed even if
 they are only zeros to run campaign-only vaccination scenarios.
}
\examples{
\dontrun{
create_campaign_vaccination_coverage_file (
 campaign_only_vaccination_coverage_file    = "campaign-only-bestcase.csv",
 routine_campaign_vaccination_coverage_file = "campaign-bestcase.csv"
 )
 }
}
