% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_no_vaccination_coverage_file}
\alias{create_no_vaccination_coverage_file}
\title{Create a vaccine coverage file for no-vaccination scenario}
\usage{
create_no_vaccination_coverage_file(
  no_vaccination_coverage_file,
  vaccination_coverage_file
)
}
\arguments{
\item{no_vaccination_coverage_file}{File name given to the 0\% vaccine
coverage scenarios}

\item{vaccination_coverage_file}{File used to generate vaccine coverage for
0\% scenarios}
}
\description{
A csv file for 0\% coverage vaccination scenario is created based on the csv
file of any of other vaccination scenarios. This is done because the coverage
 file downloaded from VIMC for the no-vaccination scenario is empty.
}
\examples{
\dontrun{
create_no_vaccination_coverage_file (
 no_vaccination_coverage_file = "no-vaccination.csv",
 vaccination_coverage_file    = "mcv1-default.csv"
 )
 }
}
