% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_PSA_data}
\alias{create_PSA_data}
\title{Create variable inputs for probabilistic sensitivity analysis}
\usage{
create_PSA_data(psa = 0, seed_state = 1, psadat_filename = NA)
}
\arguments{
\item{psa}{An integer indicating the total runs of PSA.}

\item{seed_state}{An integer for selecting the seed number.}

\item{psadat_filename}{A file name containing ID and input parameters of each
run of PSA.}
}
\description{
Generate latin hyper cube samples of input parameters for probabilistic
sensitivity analysis (PSA). Four variables are included in the output: (1)
"take1_input"-intercept of the regression model for age-varying vaccine
protection under 'take' assumption, (2) "take2_input"-coefficient of the
regression model for age-varying vaccine protection under 'take' assumption,
(3) "take3_input"-vaccine protection for the second dose under 'take'
assumption, (4) "mortality_input"-proportion of change in case fatality
ratios.
}
\examples{
create_PSA_data (psa = 20,
                 seed_state = 1,
                 psadat_filename = "psa_variable.csv")
}
