% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_vaccine_coverage_routine_sia}
\alias{create_vaccine_coverage_routine_sia}
\title{Create vaccine coverage files by scenarios}
\usage{
create_vaccine_coverage_routine_sia(
  vaccine_coverage_folder = "",
  coverage_prefix = "",
  touchstone = "",
  antigen = "",
  scenario_name,
  vaccine_coverage_subfolder = ""
)
}
\arguments{
\item{vaccine_coverage_folder}{A folder name for the vaccine coverage files.
Include a slash at the end.}

\item{coverage_prefix}{A prefix used in the name of vaccine coverage file.}

\item{touchstone}{A version note in the file name used by VIMC. Include a
underscore at the beginning and end of the name.}

\item{antigen}{A disease name used by VIMC: "Measles".}

\item{scenario_name}{A name of vaccination scenarios.}

\item{vaccine_coverage_subfolder}{A folder name under the \code{x} folder for
the vaccine coverage files.}
}
\description{
Generates files for routine programmes (MCV1, MCV2) and supplementary
immunisation activities (SIAs) from the vaccine coverage files.
}
\examples{
  create_vaccine_coverage_routine_sia (
  vaccine_coverage_folder    = "vaccine_coverage/",
  coverage_prefix            = "coverage",
  touchstone                 = "_201910gavi-5_",
  antigen                    = "measles-",
  scenario_name              = "campaign-only-bestcase",
  vaccine_coverage_subfolder = "scenarios/")
}
