% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_cfr}
\alias{data_cfr}
\title{CFRs}
\format{
A data table with 78477 observations of 7 variables.
\describe{
  \item{Country}{Country name}
  \item{Code}{ISO3 country code}
  \item{GBD Region}{GBD regions}
  \item{Grouping}{GBD regions further grouped by under-5 mortality rate,
  for example, LAC>=50 and LAC<50}
  \item{Year}{Year}
  \item{over5}{CFRs for the age group over 5 years old}
  \item{uner5}{CFRs for the age group under 5 years old}
}
}
\source{
{check with Kevin/Kaja/Mark}
}
\usage{
data_cfr
}
\description{
A dataset containing the case fatality ratios (CFRs) by calender year and age
 groups of under and over 5 years.
}
\keyword{datasets}
