% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_cfr_portnoy}
\alias{data_cfr_portnoy}
\title{CFRs for DALY calculation with Portnoy estimates}
\format{
A data table with 156240 observations of 45 variables.
\describe{
  \item{country_name}{Country name}
  \item{country_code}{ISO3 country code}
  \item{GBD Region}{GBD regions}
  \item{U5MR}{Groups for under-5 mortaltiy rate, >=50 or LAC<50}
  \item{year}{Year}
  \item{cfr_over5_scenario_s4}{CFRs for the age group over 5 years under one
  of the 10 coverage scenarios. The \code{s4} refers to declining trend after
   2018.}
  \item{cfr_over5_scenario_s6}{CFRs for the age group over 5 years under one
  of the 10 coverage scenarios. The \code{s6} refers to static trend of the
  2018 level in the future.}
  \item{cfr_uner5_scenario_s4}{CFRs for the age group under 5 years old,
  under one of the 10 coverage scenarios. The \code{s4} refers
  to declining trend after 2018.}
  \item{cfr_uner5_scenario_s6}{CFRs for the age group under 5 years old,
  under one of the 10 coverage scenarios. The \code{s6} refers to static
  trend of the 2018 level in the future.}
}
}
\source{
{Portnoy A, Jit M, Ferrari M, Hanson M, Brenzel L, Verguet S.
Estimates of case-fatality ratios of measles in low-income and middle-income
countries: a systematic review and modelling analysis. Lancet Glob Health
2019; 7: e472–81.}
}
\usage{
data_cfr_portnoy
}
\description{
A dataset containing the case fatality ratios (CFRs) by scenario, calender
year and age groups of under and over 5 years. using Portnoy estimates.
}
\keyword{datasets}
