% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_cfr_wolfson}
\alias{data_cfr_wolfson}
\title{CFRs with Wolfson estimates for calculating deaths}
\format{
A data table with 294 observations of 3 variables.
\describe{
  \item{country_code}{ISO3 country code}
  \item{Country}{Country name}
  \item{CFR}{Case fatality ratios}
}
}
\source{
{Wolfson LJ, Grais RF, Luquero FJ, Birmingham ME, Strebel PM.
Estimates of measles case fatality ratios: a comprehensive review of
community-based studies. Int J Epidemiol 2009; 38: 192--205.}
}
\usage{
data_cfr_wolfson
}
\description{
A dataset containing the country-specific case fatality ratios among children
 under 5 years old, using Wolfson estimates.
}
\keyword{datasets}
