% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_contact}
\alias{data_contact}
\title{Contact matrix by age}
\format{
A data table with 10302 observations of 102 variables.
\describe{
  \item{contact.age}{age between 0 and 100}
  \item{V1-V101}{age between 0 and 100}
}
}
\source{
{Mossong J, Hens N, Jit M, et al.
Social Contacts and Mixing Patterns Relevant to the Spread of Infectious Diseases.
PLoS Med 2008; 5(3): e74.}
}
\usage{
data_contact
}
\description{
A dataset containing the contact matrix for age between 0 and 100,
 adapted from UK population in the POLYMOD study.
}
\keyword{datasets}
