% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_contact_polymod}
\alias{data_contact_polymod}
\title{Contact matrices - POLYMOD UK extension}
\format{
A list with multiple data tables denoted by ISO3 country code. Each
data table contains 101*101 observations.
}
\source{
{Mossong J, Hens N, Jit M, et al. Social contacts and mixing patterns
 relevant to the spread of infectious diseases. PLoS Med 2008; 5(3): e74.}

\url{https://population.un.org/wpp/Download/Standard/Population/}
}
\usage{
data_contact_polymod
}
\description{
A dataset containing contact matrices of yearly age bands between 0 and 100
in 195 countries, adapted from physical contact in the Great Britain in the
POLYMOD study. Expansion of the age groups in the matrix is based on
country-specific population distribution at the reference year (2020).
}
\keyword{datasets}
