% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_contact_uk}
\alias{data_contact_uk}
\title{Contact matrix - POLYMOD UK}
\format{
A data table contains 101*101 observations, with age of contactors in
 row and age of contactees in column.
}
\source{
{Mossong J, Hens N, Jit M, et al. Social contacts and mixing patterns
 relevant to the spread of infectious diseases. PLoS Med 2008; 5(3): e74.}
}
\usage{
data_contact_uk
}
\description{
A dataset containing a contact matrix of yearly age bands between 0 and 100,
adapted from physical contact in the Great Britain in the POLYMOD study.
}
\keyword{datasets}
