% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_lexp0}
\alias{data_lexp0}
\title{Life expectancy at birth}
\format{
A data table with 137712 observations of 8 variables.
\describe{
  \item{country_code_numeric}{Numeric country code}
  \item{country_code}{ISO3 country code}
  \item{country}{Country name}
  \item{age_from}{age from (start-age)}
  \item{age_to}{age to (end-age)}
  \item{year}{Year}
  \item{gender}{Gender}
  \item{LE0}{Life expectancy at birth (\code{age_from=age_to=0})}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Population/}

{VIMC}
}
\usage{
data_lexp0
}
\description{
A dataset containing estimates of life expectancy at birth.
}
\keyword{datasets}
