% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_lexp_remain}
\alias{data_lexp_remain}
\title{Life expectancy for a specific age group (remain)}
\format{
A data table with 601920 observations of 8 variables.
\describe{
  \item{country_code_numeric}{Numeric country code}
  \item{country_code}{ISO3 country code}
  \item{country}{Country name}
  \item{age_from}{Age from (start-age)}
  \item{age_to}{Age to (end-age)}
  \item{year}{Year}
  \item{gender}{Gender}
  \item{value}{Life expectacy for the age group between age_from and age_to}
}
}
\source{
\url{https://population.un.org/wpp/Download/Standard/Population/}

{VIMC}
}
\usage{
data_lexp_remain
}
\description{
A dataset containing estimates of life expectancy for a specific age group.
}
\keyword{datasets}
