% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_r0}
\alias{data_r0}
\title{R0 by country}
\format{
A data table with 336 observations of 3 variables.
\describe{
  \item{contact.age}{age between 0 and 100}
  \item{V1:V101}{age between 0 and 100}
}
}
\source{
{Estimates provided by and Ferrari and team}
}
\usage{
data_r0
}
\description{
A dataset containing the R0 of measles in 112 countries.
}
\keyword{datasets}
