% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_template}
\alias{data_template}
\title{template for outputs}
\format{
A data table with 8908200 observations of 9 variables.
\describe{
  \item{disease}{Disease (="Measles")}
  \item{year}{Year, from 2000 to 2100}
  \item{age}{Age, from 0 to 100 years old}
  \item{country}{ISO3 country code}
  \item{country_name}{Country name}
  \item{cohort_size}{Population size}
  \item{deaths}{Disease-specific deaths}
  \item{cases}{Disease-specific cases}
  \item{dalys}{Disease-specific disability-adjusted life years}
}
}
\source{
{VIMC}
}
\usage{
data_template
}
\description{
A template dataset containing countries, years, ages, and other outputs.
}
\keyword{datasets}
