% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_timeliness}
\alias{data_timeliness}
\title{Timeliness data}
\format{
A data table with 34496 observations of 4 variables.
\describe{
  \item{country_code}{ISO3 country code}
  \item{age}{age in weeks. \code{NA} represents the overall population. For
  0-1 year old (0-52 weeks),  \code{timeliness} is presented each week; for
  1-3 year old (53-156 weeks), it is presented the first week of each month.}
  \item{timeliness}{timeliness coverage according to age}
  \item{prop_final_cov}{final coverage of the population. Exist only when
  age is \code{NA}}
}
}
\source{
{Clark, A and Sanderson, C. Timing of children's vaccinations in 45
low-income and middle-income countries: an analysis of survey data. Lancet
2009; 373(9674): 1543-1549.}

{Check with Kevin/Mark for how the csv is processed}
}
\usage{
data_timeliness
}
\description{
A dataset containing the timeliness data estimates and final coverage.
}
\keyword{datasets}
