% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dxplot.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{diagnostic_plots}
\usage{
diagnostic_plots(
  vaccine_coverage_folder,
  coverage_prefix,
  touchstone,
  antigen,
  scenarios,
  base_scenario,
  burden_estimate_folder,
  plot_folder,
  group_name,
  countries,
  cfr_options,
  psa,
  start_year = -1,
  end_year = -1,
  compare_plots = FALSE
)
}
\arguments{
\item{vaccine_coverage_folder}{Folder name for the vaccine coverage files.
Include a slash in the end.}

\item{coverage_prefix}{Prefix of vaccine coverage file names.}

\item{touchstone}{Version note used by VIMC. Include a underscore at the
beginning and the end.}

\item{antigen}{Disease name used by VIMC.}

\item{scenarios}{Name of vaccination strategies included for evaluation.}

\item{base_scenario}{Name of the vaccination strategy that is used as the
baseline for calculating averted burden.}

\item{burden_estimate_folder}{A folder name for the file which contains the
model outputs for evaluation.Include a slash at the end.}

\item{plot_folder}{A folder name for the diagnostic plots to be saved.}

\item{group_name}{A modelling group name used by VIMC.}

\item{countries}{A vector of ISO country codes for plotting.}

\item{psa}{Number of runs for probabilistic sensitivity analysis.}

\item{start_year}{Beginning year of the plot.}

\item{end_year}{End year of the plot.}

\item{compare_plots}{A logical variable to determine whether to combine model
outputs across different scenarios and compare them in the same plot.}

\item{cfr_option}{The method used for specifying the trend of case fatality
ratios (CFRs): "Wolfson" or/and "Portnoy".}
}
\description{
Create diagnostic plots of vaccine coverages and burden estimates (cases,
deaths, dalys) for selected countries.
}
\examples{
diagnostic_plots (
  vaccine_coverage_folder    = "vaccine_coverage/",
  coverage_prefix            = "coverage",
  touchstone                 = "_201910gavi-5_",
  antigen                    = var$antigen,
  scenarios                  = scenarios[1:3],
  base_scenario              = base_scenario,
  burden_estimate_folder     = "central_burden_estimate/",
  plot_folder                = "plots/",
  group_name                 = "LSHTM-Jit-",
  countries                  = c("BGD","ETH"),
  cfr_options                = c("Wolfson", "Portnoy"),
  psa                        = 200,
  start_year                 = 2000,
  end_year                   = 2100,
  compare_plots              = FALSE)
}
