% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_deaths_dalys}
\alias{estimate_deaths_dalys}
\title{Estimate deaths and DALYs of measles}
\usage{
estimate_deaths_dalys(
  cfr_option,
  burden_estimate_file,
  burden_estimate_folder,
  vimc_scenario,
  portnoy_scenario,
  psa = 0
)
}
\arguments{
\item{burden_estimate_file}{File name for model outputs to be saved.}

\item{burden_estimate_folder}{Folder name where \code{burden_estimate_file} is
saved. Include a forward slash at the end.}

\item{vimc_scenario}{Name for vaccine strategy used in \code{data_cfr_portnoy.rda},
for extracting CFR estimates by the Portnoy method.}

\item{portnoy_scenario}{Scenario name for future CFR trends used in the
Portnoy method: "s4" - declining after 2018, or "s6" - staying stagnant at
the 2018 level.}

\item{psa}{A numeric variable of the total runs for PSA. Use 0 to indicate a
single run without PSA (central estimate).}

\item{cfr_options}{Methods for CFR estimates: "Wolfson" or/and "Portnoy".}
}
\description{
A function that estimates the number of deaths by applying case fatality
rates (CFRs) to the number of cases, and then calculates the
disability-adjusted years (DALYs)
}
\examples{
\dontrun{
estimate_deaths_dalys (
 cfr_option             = "Portnoy",
 burden_estimate_file   = burden_estimate_file,
 burden_estimate_folder = "central_burden_estimate/",
 vimc_scenario          = "campaign-only",
 portnoy_scenario       = "s6",
 psa                    = 0
 )
 }
}
