% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandMatrix}
\alias{expandMatrix}
\title{Expand the matrix to a different dimension
This function returns an expanded contact matrix with the specified age
structure for model inputs.}
\usage{
expandMatrix(
  A,
  expand_rows = 1,
  expand_cols = 1,
  rescale_rows = F,
  rescale_cols = F
)
}
\arguments{
\item{A}{A matrix to be expanded.}

\item{expand_rows}{A number of times to repeat each row.}

\item{expand_cols}{A number of times to repeat each column.}

\item{rescale_rows}{A logical variable to control whether to rescale the
expanded rows.}

\item{rescale_cols}{A logical variable to control whether to rescale the
expanded columns.}
}
\value{
An expanded matrix with rescaling if applicable.
}
\description{
Expand the matrix to a different dimension
This function returns an expanded contact matrix with the specified age
structure for model inputs.
}
\examples{
expandMatrix (matrix(1:9,3,3), 2, 1, FALSE, FALSE)
}
