% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runCountry}
\alias{runCountry}
\title{Execute the fortran model for a country and a PSA run}
\usage{
runCountry(
  ii,
  iso3,
  years,
  vaccination,
  using_sia,
  dinf,
  gamma_rate,
  r0_basic,
  amplitude,
  take,
  degree,
  sia.method,
  coverage_routine,
  coverage_sia,
  timeliness,
  contact,
  c_rnought,
  population,
  tstep,
  save.scenario,
  foldername,
  measles_model,
  debug_model,
  debug_spinup,
  debug_age,
  debug_compartments,
  debug_country,
  debug_relative,
  debug_timepoints,
  r,
  runs,
  psa,
  psa_var,
  run_model,
  remove_files,
  contact_mat
)
}
\arguments{
\item{ii}{Numeric order of the selected country for evaluation.}

\item{iso3}{ISO-3 code of the selected country.}

\item{years}{A vector containing continuous calender years for simulation.}

\item{vaccination}{A Numeric indicator that determines vaccination programmes
for children: 0 - No vaccination, 1 - Only MCV1,  2 - MCV1 and MCV2.}

\item{using_sia}{A numeric indicator that determines Whether supplementary
immunisation activities (SIAs) are implemented: 0 - no SIA, 1 - with SIA.}

\item{dinf}{Duration of infection in days.}

\item{gamma_rate}{Recovery rate, with a unit of 1/(\code{dinf}*\code{tstep}*year).}

\item{r0_basic}{Basic reproduction number (R0) calculated by the eigenvalue
of the product of contact matrix (\code{contact}) and duration of infection
(\code{dinf}).}

\item{amplitude}{A numeric variable for amplitude for seasonality.}

\item{take}{A 1*3 vector that denotes the level of vaccine protection using
the 'take' ('all-or-not') assumption. Vaccine protection is assumed to vary
by age under a linear regression, with (1) and (2) elements of the vector
representing the intercept and coefficient of the first dose, respectively.
(3) element represents the vaccine protection by the second dose. Note that
in the previous fortran model without the incorporation of age-related
regression model, (1) and (2) elements were directly indicating protection of
 the first dose before and after 1 year old, respectively.}

\item{degree}{A 1*3 vector that denotes the level of vaccine protection using
 the 'degree' ('leak') assumption. (1) and (2) elements represents the
 protection of the first dose before and after 1 year old, respectively, and
(3) element shows that of the second dose. Note that the second dose only has
 an effect if \code{vaccine=2}.}

\item{sia.method}{A Numeric indicator that determines how SIAs are carried
out: 1 - varying \code{take}, 2 - varying \code{degree}.}

\item{coverage_routine}{A data frame for routine vaccination coverage under
a selected scenario.}

\item{coverage_sia}{A data frame for SIA coverage under a selected scenario.}

\item{timeliness}{A data frame for timeliness estimates by age.}

\item{contact}{A data frame for the contact matrix by age (0-100 years old).}

\item{c_rnought}{A numeric variable of R0 value for the selected country
\code{iso3}. Estimates provided by Ferrari and colleagues.}

\item{population}{A data frame for population size by country and age.}

\item{tstep}{Simulation time steps for each year.}

\item{save.scenario}{A folder name of the selected scenario.}

\item{foldername}{A folder name for input and output files of the selected
scenario. It typically does not exist but may come in handy when only
processing results.}

\item{measles_model}{An executable file that processes fortan codes of the
measles model.}

\item{debug_model}{A logical variable that determines whether to debug the
model.}

\item{debug_spinup}{A logical variable that determines whether to generate
outputs for the spin-up (equilibrium) period.}

\item{debug_age}{A numeric variable that determines the format of output age
groups: 0 - all in annual age-strata, 1 - age between 0 and 2 in weekly
age-strata and age between 3 and 100 in annual age-strata.}

\item{debug_country}{A vector including ISO3 codes of country to debug, or *
to debug all countries.}

\item{debug_relative}{A logical variable that determines the output format of
new cases: TRUE - proportion, and FALSE - absolute number.}

\item{debug_timepoints}{A logical variable that determines the output unit of
time: 0 - per year, 1 - per timepoint and report first 25% of timepoints, and
2 - per timepoint and report all timepoints.}

\item{r}{A numeric variable of the order of the PSA runs.}

\item{runs}{A numeric variable of the total runs for each scenario.}

\item{psa}{A numeric variable of the total runs for PSA. Use 0 to indicate a
single run without PSA.}

\item{psa_var}{A data frame including the parameter values used in PSA.}

\item{run_model}{A logical variable that determines whether to run the
measles model.}

\item{remove_files}{A logical variable that determines whether to remove
output files after finishing a single country run.}

\item{contact_mat}{A character variable that indicates the assumption for
contact pattern. "nomix" - homogeneous mixing, "uk" - POLYMOD Great Britain
physical contacts, and "syn" - country-specific synthetic matrix.}

\item{debug_compartment}{A numeric variable that determines the compartments
to output: 0 - number of cases, and 1 - each compartment.}
}
\description{
A function nested under \code{\link{runScenario}} to run the fortran codes
 for measles vaccination, given a particular country and a variable set of
 probabilistic sensitivity analysis (PSA).
}
\examples{
runCountry (
  ii                 = 2,
  iso3               = "BGD",
  years              = 1980:2020,
  vaccination        = 1,
  using_sia          = 1,
  dinf               = 14,
  gamma_rate         = 0.02607,
  r0_basic           = 13,
  amplitude          = 0.05,
  take               = c(0.64598, 0.01485, 0.98),
  degree             = c(0.85, 0.95, 0.98),
  sia.method         = 1,
  coverage_routine   = coverage_routine,
  coverage_sia       = coverage_sia,
  timeliness         = timeliness,
  contact            = contact,
  c_rnought          = 10,
  population         = population,
  tstep              = 1000,
  save.scenario      = "scenario01",
  foldername         = NULL,
  measles_model      = "vaccine2019_sia_singlematrix.exe",
  debug_model        = FALSE,
  debug_spinup       = FALSE,
  debug_age          = 0,
  debug_compartments = 1,
  debug_country      = "*",
  debug_relative     = FALSE,
  debug_timepoints   = 0,
  r                  = 1,
  runs               = 1,
  psa                = 0,
  psa_var            = psa_var,
  run_model          = TRUE,
  remove_files       = FALSE,
  contact_mat        = "nomix")
}
