% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runScenario}
\alias{runScenario}
\title{Run the measles model for a selected vaccination scenario}
\usage{
runScenario(
  vaccine_coverage_folder = "",
  coverage_prefix = "",
  touchstone = "",
  antigen = "",
  scenario_name,
  scenario_number,
  vaccine_coverage_subfolder = "",
  burden_estimate_folder,
  group_name,
  countries = "all",
  cluster_cores = 1,
  psa = 0,
  vaccination,
  using_sia,
  measles_model,
  debug_model = FALSE,
  fix.uk.contact = TRUE
)
}
\arguments{
\item{vaccine_coverage_folder}{A folder name for the vaccine coverage files.}

\item{coverage_prefix}{A prefix used in the name of vaccine coverage file.}

\item{touchstone}{A version note in the file name used by VIMC. Include a
underscore at the beginning and end of the name.}

\item{antigen}{A disease name used by VIMC: "Measles".}

\item{scenario_name}{A name of vaccination scenarios.}

\item{scenario_number}{A folder name of the selected scenario.}

\item{vaccine_coverage_subfolder}{A folder name under the \code{x} folder for
the vaccine coverage files.}

\item{burden_estimate_folder}{A folder name for the file which contains the
model outputs for evaluation. Include a slash at the end.}

\item{group_name}{A modelling group name used by VIMC.}

\item{countries}{A vector of ISO-3 country codes used in the analysis. Use
"all" to include all countries.}

\item{cluster_cores}{A number of cores to be used in the cluster.}

\item{psa}{A numeric variable of the total runs for PSA. Use 0 to indicate a
single run without PSA.}

\item{vaccination}{A numeric indicator that determines vaccination programmes
for children: 0 - No vaccination, 1 - Only MCV1,  2 - MCV1 and MCV2.}

\item{using_sia}{A numeric indicator that determines Whether supplementary
immunisation activities (SIAs) are implemented: 0 - no SIA, 1 - with SIA.}

\item{measles_model}{An executable file that processes fortan codes of the
measles model.}

\item{debug_model}{A logical variable that determines whether to debug the
model.}
}
\description{
A function that execute the measles model under a selected vaccination
scenario, including a pre-specified set of countries and runs of probability
sensitivity analysis (PSA).
}
\examples{
runScenario (
  vaccine_coverage_folder    = "vaccine_coverage/",
  coverage_prefix            = "coverage",
  touchstone                 = "_201910gavi-5_",
  antigen                    = "measles-",
  scenario_name              = "campaign-only-bestcase",
  scenario_number            = scenario_number,
  vaccine_coverage_subfolder = "scenarios/"
  burden_template            = var$burden_template,
  burden_estimate_folder     = "central_burden_estimate/",
  group_name                 = "LSHTM-Jit-",
  countries                  = c("BGD","ETH"),
  cluster_cores              = 3,
  psa                        = 0,
  vaccination                = 0,
  using_sia                  = 1,
  measles_model              = "vaccine2019_sia_singlematrix.exe",
  debug_model                = FALSE)
}
