% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tailor_data_cfr}
\alias{tailor_data_cfr}
\title{Tailor the data structure for case fatality ratios (CFRs)}
\usage{
tailor_data_cfr(sel_countries)
}
\arguments{
\item{sel_countries}{ISO-3 codes for countries included for evaluation. If
"all", all countries in the original data are selected.}
}
\description{
Tailor \code{\link{data_cfr}} to the format for processing burden estimates.
 Changes include calender year and age structures.
}
\examples{
cfr.year.all <- tailor_data_cfr (sel_countries = "all")
}
