% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{updateProgress}
\alias{updateProgress}
\title{updateProgress (temporarily disabled)}
\usage{
updateProgress(country, cnt, runs, r, status)
}
\arguments{
\item{country}{Name of country.}

\item{cnt}{Numeric order of countries.}

\item{runs}{Total number of runs.}

\item{r}{A numeric order of the current run.}

\item{status}{A numeric variable for the stage of simulation progress: \code{1}
-Generate the input data, \code{2}-Start to run the fortran code, \code{3}-
Finish the fortran code, \code{4}-Return the processed results.}
}
\description{
update the log file to show progress
}
\examples{
updateProgress ('BGD', 1, 200, 25, 1)
}
