% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gavi_log.R
\name{writelog}
\alias{writelog}
\title{writelog}
\usage{
writelog(logname, x)
}
\arguments{
\item{logname}{Name of log file.}

\item{x}{Text content to be added.}
}
\description{
Write a text message to a log file with the current time tag.
}
\examples{
writelog ("gavi_log", "Start generating data")
}
