% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_contact_syn}
\alias{data_contact_syn}
\title{Contact matrices - Synthetic}
\format{
A list with multiple data tables denoted by ISO3 country code. Each
data table contains 101*101 observations.
}
\source{
\url{http://github.com/kieshaprem/synthetic-contact-matrices/}

{Prem K, Cook AR, Jit M. Projecting social contact matrices in 152
countries using contact surveys and demographic data. PLoS Comput Biol 2017;
13(9): e1005697.}
}
\usage{
data_contact_syn
}
\description{
A dataset containing synthetic contact matrices of yearly age bands between 0
 and 100 in 177 countries.
}
\keyword{datasets}
