% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_r0}
\alias{data_r0}
\title{R0 by country}
\format{
A data table with 336 observations of 3 variables.
\describe{
  \item{country}{Country name}
  \item{country_code}{ISO3 country code}
  \item{r0}{Numeric value of R0 estimate}
}
}
\source{
{Estimates provided by and Ferrari M and team}
}
\usage{
data_r0
}
\description{
A dataset containing R0 estimates of measles in 112 countries.
}
\keyword{datasets}
