% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimate_deaths_dalys}
\alias{estimate_deaths_dalys}
\title{Estimate deaths and DALYs of measles}
\usage{
estimate_deaths_dalys(
  cfr_option,
  burden_estimate_file,
  burden_estimate_folder,
  vimc_scenario,
  portnoy_scenario,
  psa = 0
)
}
\arguments{
\item{burden_estimate_file}{A file name for model outputs to be saved.}

\item{burden_estimate_folder}{A folder name for the file which contains the
model outputs for evaluation. Include a slash at the end.}

\item{vimc_scenario}{A scenario name for vaccine coverage used in
\code{data_cfr_portnoy.rda}, for extracting CFR estiamtes by the Portnoy
method.}

\item{portnoy_scenario}{A scenario name for future CFR trends when using the
Portnoy method: "s4" - declining after 2018, or "s6" - staying stagnant at
the 2018 level.}

\item{psa}{A numeric variable of the total runs for PSA. Use 0 to indicate a
single run without PSA.}

\item{cfr_options}{Methods for CFR estimates: "Wolfson" or/and "Portnoy".}
}
\description{
A function that estimates the number of deaths by applying case fatality
rates (CFRs) to the number of cases, and then calculates the
disability-adjusted years (DALYs)
}
\examples{
estimate_deaths_dalys (
 cfr_option             = "Portnoy",
 burden_estimate_file   = burden_estimate_file,
 burden_estimate_folder = "central_burden_estimate/",
 vimc_scenario          = "campaign-only",
 portnoy_scenario       = "s6",
 psa                    = 0
 )
}
