% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runScenario}
\alias{runScenario}
\title{Run the measles model for a selected vaccination scenario}
\usage{
runScenario(
  vaccine_coverage_folder = "",
  vaccine_coverage_subfolder = "",
  coverage_prefix = "",
  antigen = "",
  scenario_name,
  save_scenario,
  burden_estimate_folder,
  group_name,
  countries = "all",
  cluster_cores = 1,
  psa = 0,
  vaccination,
  using_sia,
  measles_model,
  debug_model = FALSE,
  contact_mat,
  step_ve,
  sim_years = 1980:2100
)
}
\arguments{
\item{vaccine_coverage_folder}{A folder name for the vaccine coverage files.}

\item{vaccine_coverage_subfolder}{A folder name under the \code{x} folder for
the vaccine coverage files.}

\item{coverage_prefix}{A prefix used in the name of vaccine coverage file.}

\item{antigen}{A disease name used by VIMC: "Measles".}

\item{scenario_name}{A name of vaccination scenarios.}

\item{save_scenario}{A folder name for saving a selected scenario.}

\item{burden_estimate_folder}{A folder name for the file which contains the
model outputs for evaluation. Include a slash at the end.}

\item{group_name}{A modelling group name used by VIMC.}

\item{countries}{A vector of ISO-3 country codes used in the analysis. Use
"all" to include all countries.}

\item{cluster_cores}{A number of cores to be used in the cluster.}

\item{psa}{A numeric variable of the total runs for PSA. Use 0 to indicate a
single run without PSA.}

\item{vaccination}{A numeric indicator that determines vaccination programmes
for children: 0 - No vaccination, 1 - Only MCV1,  2 - MCV1 and MCV2.}

\item{using_sia}{A numeric indicator that determines Whether supplementary
immunisation activities (SIAs) are implemented: 0 - no SIA, 1 - with SIA.}

\item{measles_model}{An executable file that processes Fortran codes of the
measles model.}

\item{debug_model}{A logical variable that determines whether to debug the
model.}

\item{contact_mat}{A character variable that indicates the assumption for
contact pattern. "prpmix" - proportional mixing informed by age distribution
of population, "unimix" -  uniform mixing (equal mixing probability across
each age year), "polymod" - POLYMOD Great Britain physical contacts expanded
using local populations, and "syn" - country-specific synthetic matrix.}

\item{step_ve}{A logical variable that determines whether a step change on
age-related vaccine efficacy is applied. This implies different meaning of
vaccine efficacy vector (\code{take}) and a corresponding \code{measles_model}
should be applied.}

\item{sim_years}{A numeric vector containing calendar years included for model
simulation.}
}
\description{
A function that execute the measles model under a selected vaccination
scenario, including a pre-specified set of countries and runs of probability
sensitivity analysis (PSA).
}
\examples{
runScenario (
  vaccine_coverage_folder    = "vaccine_coverage_upd/",
  vaccine_coverage_subfolder = "scenarios/"
  coverage_prefix            = "coverage",
  antigen                    = "measles-",
  scenario_name              = "campaign-only-default",
  save_scenario               = scenario_number,
  burden_estimate_folder     = "central_burden_estimate/",
  group_name                 = "LSHTM-Jit-",
  countries                  = c("BGD","ETH"),
  cluster_cores              = 3,
  psa                        = 0,
  vaccination                = 0,
  using_sia                  = 1,
  measles_model              = "vaccine2019_sia_singlematrix.exe",
  debug_model                = FALSE,
  contact_mat                = "prpmix",
  step_ve                    = FALSE,
  sim_years                  = 1980:2100
  )
}
