% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tailor_data_lexp_remain}
\alias{tailor_data_lexp_remain}
\title{Tailor the data structure for life expectancy by age and year}
\usage{
tailor_data_lexp_remain(sel_countries = "all")
}
\arguments{
\item{sel_countries}{ISO-3 codes for countries included for evaluation. If
"all", all countries in the original data are selected.}
}
\description{
Tailor the \code{\link{data_lexp_remain}} data to the format for processing burden
estimates. Linear interpolation between calender years was applied.
}
\examples{
lexp_remain <- tailor_data_lexp_remain (sel_countries = c("AGO","BGD"))
}
